Bert_LCF_ATEPC_ABSA/requirements.txt
pytorch~=1.12.1
cudatoolkit~=11.3.1
tqdm~=4.64.1
transformers~=4.24.0
pytorch-crf~=0.7.2
pandas~=1.5.1
matplotlib~=3.5.1
scikit-learn~=1.1.1